Update CONTRIBUTING.md
There is no text on line 566 (also the lines are different between this one and the one in my VS code, which is odd, they are about 30ish apart). However, on line 566 there is a grammatical error:
line was - Alternatively it possible to run pre-commit via tox.  

I changed it to- Alternatively it is possible to run pre-commit via tox. 

Is that the one you had planted in?